# cli: call gc.freeze before awaiting applet task