I noticed that while an applet is running, a full garbage collection takes ~20ms. This 20ms delay means that an applet that either sends packets/fills transfers faster than one per ~1.25ms potentially overflows. The 20ms delay is almost entirely due to the garbage collector tracking objects that exist prior to the applet running. Additionally, most of these objects do not become collectable until after the applet has finished running. Calling `gc.collect()` followed by `gc.freeze()` immediately before awaiting the applet task reduces the full gc delay to ~>1ms. I think the only downside to this addition is an additional ~20ms of delay for starting an applet.

Below are some snippets of code I used to make measurements.

By running the following code immediatly before awaiting the applet task:
```
gc.set_debug(gc.DEBUG_STATS)
gc.collect()
gc.collect()
```
We can see that there are ~60000 objects tracked in generation 2 before awiting the applet task.
```
gc: collecting generation 2...
gc: objects in each generation: 123 0 84349
gc: objects in permanent generation: 375
gc: done, 20757 unreachable, 0 uncollectable, 0.0266s elapsed
gc: collecting generation 2...
gc: objects in each generation: 0 0 63705
gc: objects in permanent generation: 375
gc: done, 0 unreachable, 0 uncollectable, 0.0198s elapsed
```
Running a latency benchmark with a large enough count to trigger a full gc:
Before change:
```
> glasgow run benchmark latency -c 25000000
I: g.device.hardware: device already has bitstream ID 397c73a1d60a12e55826056425ca030c
I: g.cli: running handler for applet 'benchmark'
I: g.applet.internal.benchmark: running benchmark mode latency for 23.842 MiB
I: g.applet.internal.benchmark: mode latency: mean: 313.93 µs stddev: 149.42 µs worst: 20640.14 µs
```
After change:
```
> glasgow run benchmark latency -c 25000000
I: g.device.hardware: device already has bitstream ID 397c73a1d60a12e55826056425ca030c
I: g.cli: running handler for applet 'benchmark'
I: g.applet.internal.benchmark: running benchmark mode latency for 23.842 MiB
I: g.applet.internal.benchmark: mode latency: mean: 292.42 µs stddev: 64.88 µs worst: 1439.93 µs
```
Checking if the frozen objects would have become collectable sometime while an applet was running
```
gc.collect()
gc.freeze()

done, pending = await asyncio.wait(tasks, return_when=asyncio.FIRST_COMPLETED)
... 
await device.demultiplexer.cancel()

gc.disable()
gc.set_debug(gc.DEBUG_STATS)
gc.collect()
gc.unfreeze()
stats_before = gc.get_stats()
gc.collect()
stats_after = gc.get_stats()
print(f"{stats_before[2]=}")
print(f"{stats_after[2]=}")
```

```
gc: collecting generation 2...
gc: objects in each generation: 398 544 776
gc: objects in permanent generation: 64057
gc: done, 1656 unreachable, 0 uncollectable, 0.0004s elapsed
gc: collecting generation 2...
gc: objects in each generation: 1 0 64110
gc: objects in permanent generation: 0
gc: done, 0 unreachable, 0 uncollectable, 0.0223s elapsed
stats_before[2]={'collections': 3, 'collected': 27063, 'uncollectable': 0}
stats_after[2]={'collections': 4, 'collected': 27063, 'uncollectable': 0}
```