@whitequark 2025-03-01
Could you test this on PyPy as well, please?