@Sazzach 2025-03-01
Oh shoot, PyPy doesn't have gc.freeze. I'll take a look and see how long it takes on PyPy, and check if something similar is possible there. At a minimum I could add a check for the interpreter version.