@Sazzach 2025-03-18
(I'm sorry I think I messed up when I tried to rebase my branch)

I didn't find an equivalent to gc.freeze() in PyPy, so I added a check for the implementation. My benchmark above was misleading, the improvement is more along the lines of ~22ms => 8ms, which I still think is valuable.